g_file_get_contents() takes gsize, not int.

+++ test/dump.c
@@ -19,7 +19,7 @@ void dump_sprite(SwfdecSprite *s);
 int main (int argc, char *argv[])
 {
   char *contents;
-  int length;
+  gsize length;
   int ret;
   char *fn = "it.swf";
   SwfdecDecoder *s;
